+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Version 4.2.1 (Under active development)
 
+### App Center
+
+* **[Feature]** Improved `AES` token encryption algorithm using `Encrypt-then-MAC` data authentication approach.
+
 ### App Center Distribute
 
 * **[Fix]** Fix a rare deadlock case when a new version starts downloading and at the same moment the download status is checked.